# Tugas/Teknologi Integrasi Sistem/UAP_TIS_TIC/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WowoClean API - Sistem Manajemen Kontainer Limbah B3",
        "description": "Enterprise REST API untuk manajemen kontainer limbah B3 dengan autentikasi JWT, otorisasi berbasis Role, dan API Gateway pattern.",
        "contact": {
            "name": "WowoClean API Support",
            "email": "admin@wowoclean.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/gateway/containers": {
            "get": {
                "tags": [
                    "Gateway - Containers"
                ],
                "summary": "Ambil daftar semua kontainer",
                "description": "Mengambil semua kontainer dengan fitur pencarian dan filter. Dapat diakses oleh semua role (admin & user).",
                "operationId": "getContainers",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pencarian berdasarkan kode kontainer atau lokasi",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "WC-LC"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter berdasarkan tipe limbah",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "limbah_cair",
                                "limbah_padat",
                                "limbah_gas"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status kontainer",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Active",
                                "Maintenance",
                                "Archived",
                                "Full"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kontainer berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data kontainer berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gateway - Containers"
                ],
                "summary": "Buat kontainer baru (Admin only)",
                "description": "Membuat kontainer limbah B3 baru. Hanya dapat diakses oleh role admin.",
                "operationId": "createContainer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "container_code",
                                    "type",
                                    "capacity",
                                    "location"
                                ],
                                "properties": {
                                    "container_code": {
                                        "type": "string",
                                        "example": "WC-LC-009"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "limbah_cair",
                                        "enum": [
                                            "limbah_cair",
                                            "limbah_padat",
                                            "limbah_gas"
                                        ]
                                    },
                                    "capacity": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "current_fill_level": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Gudang Baru - Surabaya"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Active",
                                        "enum": [
                                            "Active",
                                            "Maintenance",
                                            "Archived",
                                            "Full"
                                        ]
                                    },
                                    "last_maintenance_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kontainer berhasil dibuat"
                    },
                    "403": {
                        "description": "Forbidden - Hanya admin"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gateway/containers/{id}": {
            "get": {
                "tags": [
                    "Gateway - Containers"
                ],
                "summary": "Ambil detail kontainer berdasarkan ID",
                "description": "Mengambil detail kontainer beserta tracking logs. Dapat diakses oleh semua role.",
                "operationId": "getContainerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kontainer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail kontainer berhasil diambil"
                    },
                    "404": {
                        "description": "Kontainer tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Gateway - Containers"
                ],
                "summary": "Update kontainer (Admin only)",
                "description": "Memperbarui data kontainer limbah B3. Hanya dapat diakses oleh role admin.",
                "operationId": "updateContainer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kontainer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "container_code": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "limbah_cair",
                                            "limbah_padat",
                                            "limbah_gas"
                                        ]
                                    },
                                    "capacity": {
                                        "type": "number"
                                    },
                                    "current_fill_level": {
                                        "type": "number"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Active",
                                            "Maintenance",
                                            "Archived",
                                            "Full"
                                        ]
                                    },
                                    "last_maintenance_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kontainer berhasil diperbarui"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Kontainer tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Gateway - Containers"
                ],
                "summary": "Hapus kontainer (Admin only)",
                "description": "Menghapus kontainer dari database. Hanya admin.",
                "operationId": "deleteContainer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kontainer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kontainer berhasil dihapus"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Kontainer tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gateway/containers/{id}/archive": {
            "patch": {
                "tags": [
                    "Gateway - Containers"
                ],
                "summary": "Archive kontainer (Admin only)",
                "description": "Mengubah status kontainer menjadi Archived. Hanya admin.",
                "operationId": "archiveContainer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kontainer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kontainer berhasil di-archive"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Kontainer tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan JWT token yang didapat dari endpoint login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}